+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/domain/validation/ValidationResult.kt
@@ -1,0 +1,6 @@
+package com.gmail.matikano9.todoapp.domain.validation
+
+data class ValidationResult(
+    val successful: Boolean,
+    val errorMessage: String? = null,
+)