+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/domain/validation/Validations.kt
@@ -1,0 +1,10 @@
+package com.gmail.matikano9.todoapp.domain.validation
+
+import javax.inject.Inject
+
+data class Validations @Inject constructor(
+    val validateTitle: ValidateTitle,
+    val validateDescription: ValidateDescription,
+    val validateDate: ValidateDate,
+    val validateTime: ValidateTime
+)