+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/presentation/todotask/ToDoTaskEvent.kt
@@ -22,6 +22,7 @@ sealed class ToDoTaskEvent {
     data class OnPriorityChanged(val priority: Priority): ToDoTaskEvent()
     data class OnTypeChanged(val type: TaskType): ToDoTaskEvent()
 
+
     data class OnDueDateChanged(
         val dueDate: LocalDate,
         val dueDateString: String = dueDate.toDateString()